Library updated to 1.070/2.055

+++ d/druntime/rt/mars.h
@@ -12,7 +12,6 @@
  *          http://www.boost.org/LICENSE_1_0.txt)
  */
 #include <stddef.h>
-#include <sys/types.h>
 
 #if __cplusplus
 extern "C" {
@@ -31,7 +30,7 @@ typedef struct Interface
 {
     struct ClassInfo *classinfo;
     struct Vtbl vtbl;
-    ptrdiff_t offset;
+    int offset;
 } Interface;
 
 typedef struct Object
