+++ d/phobos/internal/object.d
@@ -265,7 +265,7 @@ struct Interface
 {
     ClassInfo classinfo;        /// .classinfo for this interface (not for containing class)
     void *[] vtbl;
-    ptrdiff_t offset;           /// offset to Interface 'this' from Object 'this'
+    int offset;                 /// offset to Interface 'this' from Object 'this'
 }
 
 import std.moduleinit;
